requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-WTF
WTForms
SQLAlchemy
bcrypt
Pillow
